Server dosen't dependes on PyQt5 now
