+++ lab/server/server.py
@@ -5,69 +5,12 @@ import threading
 import tornado.ioloop
 import tornado.web
 import tornado.websocket
-from PyQt5 import QtCore, QtWidgets
 from tornado import httpserver
 
 from ..config import config as cfg
 from ..device.client import InstrumentManager
 from ..device.protocol import DEFAULT_PORT, Transport
 from .handlers import *
-
-
-class ActionGroupBox(QtWidgets.QGroupBox):
-    def __init__(self, parent=None):
-        super(ActionGroupBox, self).__init__('Actions', parent=parent)
-        layout = QtWidgets.QHBoxLayout(self)
-
-        self.start_btn = QtWidgets.QPushButton("Start", parent=self)
-        self.stop_btn = QtWidgets.QPushButton("Stop", parent=self)
-        self.stop_btn.setEnabled(False)
-
-        layout.addWidget(self.start_btn)
-        layout.addWidget(self.stop_btn)
-
-        self.start_btn.clicked.connect(self.start)
-        self.stop_btn.clicked.connect(self.stop)
-
-    @QtCore.pyqtSlot()
-    def start(self):
-        self.start_btn.setEnabled(False)
-        self.stop_btn.setEnabled(True)
-
-    @QtCore.pyqtSlot()
-    def stop(self):
-        self.start_btn.setEnabled(True)
-        self.stop_btn.setEnabled(False)
-
-
-class InstruentServerMainWindow(QtWidgets.QMainWindow):
-    on_start = QtCore.pyqtSignal()
-    on_stop = QtCore.pyqtSignal()
-
-    def __init__(self, server):
-        super(InstruentServerMainWindow, self).__init__()
-        self.initUI()
-        self.server = server
-
-    def initUI(self):
-        self.centralwidget = QtWidgets.QWidget(self)
-        self.layout = QtWidgets.QVBoxLayout(self.centralwidget)
-        self.actions = ActionGroupBox()
-        self.layout.addWidget(self.actions)
-        self.setCentralWidget(self.centralwidget)
-        self.setWindowTitle(u'Hello (QuLab v0.1)')
-        self.actions.start_btn.clicked.connect(self.start)
-        self.actions.stop_btn.clicked.connect(self.stop)
-
-    @QtCore.pyqtSlot()
-    def start(self):
-        self.server.start()
-        self.on_start.emit()
-
-    @QtCore.pyqtSlot()
-    def stop(self):
-        self.server.stop()
-        self.on_stop.emit()
 
 
 def make_app(config):
@@ -105,10 +48,8 @@ class InstrumentServer:
 
 
 def main():
-    import sys
     import ssl
     import os
-    from PyQt5.QtWidgets import QApplication
     from lab._bootstrap import connect_db
 
     cookie_secret = base64.urlsafe_b64encode(ssl.RAND_bytes(32)).decode()
@@ -124,13 +65,17 @@ def main():
     ssl_ctx = ssl.create_default_context(ssl.Purpose.CLIENT_AUTH)
     ssl_ctx.load_cert_chain(cfg['ssl']['cert'], cfg['ssl']['key'])
 
-    app = QApplication(sys.argv)
     connect_db()
     server = InstrumentServer(
         config=config,
         port=cfg.get('server_port', DEFAULT_PORT),
         ssl_options=ssl_ctx)
-    win = InstruentServerMainWindow(server)
-    win.show()
-    # server.run_for_ever()
-    sys.exit(app.exec_())
+    try:
+        from . import ui
+        import sys
+        app = ui.QtWidgets.QApplication(sys.argv)
+        win = ui.InstruentServerMainWindow(server)
+        win.show()
+        sys.exit(app.exec_())
+    except:
+        server.run_for_ever()
